domain => config: move SimpleLens to only module that uses it
